requirements.txt
boto3==1.26.11
botocore==1.29.11
python-dotenv==0.21.0
requests==2.28.1
